refactor(session): extract MessageProjector
SessionService owned both the session lifecycle and the entire message
data path. The two only ever shared the engine identity guard, so the
~590-line message cluster (inbound, own-send echo, acks, revokes,
reactions, edits, history backfill, status dispatch) moves to its own
service.

MessageService now takes recordOutboundMessageEdit from the projector
directly. Inbound and outbound edits still share ONE mutation queue, which
is what keeps latest-write-wins across the two entry points; verified
against the real DI container that both resolve to the same instance.

The dead KeyedMutationQueue left behind on SessionService is removed, and
the leak test now reads the projector's live queue (mutation-probed to
confirm it is not asserting on undefined).
