+++ src/modules/message/message.service.spec.ts
@@ -6,6 +6,7 @@ import { MessageService } from './message.service';
 import { Message, MessageDirection, MessageStatus } from './entities/message.entity';
 import { SessionService } from '../session/session.service';
 import { EngineRegistry } from '../../engine/engine-registry.service';
+import { MessageProjector } from '../session/message-projector.service';
 import type { IWhatsAppEngine } from '../../engine/interfaces/whatsapp-engine.interface';
 import { HookManager } from '../../core/hooks';
 import { TemplateService } from '../template/template.service';
@@ -42,6 +43,7 @@ describe('MessageService', () => {
   let repository: jest.Mocked<Partial<Repository<Message>>>;
   let sessionService: jest.Mocked<Partial<SessionService>>;
   let engines: EngineRegistry;
+  let messageProjector: { recordOutboundMessageEdit: jest.Mock };
   let hookManager: jest.Mocked<Partial<HookManager>>;
   let templateService: jest.Mocked<Partial<TemplateService>>;
   let lidMappingStore: { lidsForPhone: jest.Mock; getCached: jest.Mock };
@@ -71,8 +73,9 @@ describe('MessageService', () => {
 
     sessionService = {
       findOne: jest.fn().mockResolvedValue({ id: 'sess-1', phone: '628123456789' }),
-      recordOutboundMessageEdit: jest.fn().mockResolvedValue(undefined),
     };
+
+    messageProjector = { recordOutboundMessageEdit: jest.fn().mockResolvedValue(undefined) };
 
     engines = new EngineRegistry();
     engines.set('sess-1', mockEngine as unknown as IWhatsAppEngine);
@@ -97,6 +100,7 @@ describe('MessageService', () => {
         { provide: getRepositoryToken(Message, 'data'), useValue: repository },
         { provide: SessionService, useValue: sessionService },
         { provide: EngineRegistry, useValue: engines },
+        { provide: MessageProjector, useValue: messageProjector },
         { provide: HookManager, useValue: hookManager },
         { provide: TemplateService, useValue: templateService },
         { provide: LidMappingStoreService, useValue: lidMappingStore },
@@ -395,11 +399,12 @@ describe('MessageService', () => {
     it('honors a configured template.renderMaxChars override', async () => {
       const configService = {
         get: (key: string, fallback: unknown) => (key === 'template.renderMaxChars' ? 10 : fallback),
-      } as unknown as ConstructorParameters<typeof MessageService>[6];
+      } as unknown as ConstructorParameters<typeof MessageService>[7];
       const capped = new MessageService(
         repository as Repository<Message>,
         sessionService as unknown as SessionService,
         engines,
+        messageProjector as unknown as MessageProjector,
         hookManager as HookManager,
         templateService as unknown as TemplateService,
         lidMappingStore as unknown as LidMappingStoreService,
@@ -1288,7 +1293,7 @@ describe('MessageService', () => {
       expect(mockEngine.editMessage).toHaveBeenCalledWith('test@c.us', 'wa-msg-1', 'edited');
       // Persistence is delegated to the session's per-message mutation queue (serialized with the
       // inbound edit path) — the service no longer writes the row directly.
-      expect(sessionService.recordOutboundMessageEdit).toHaveBeenCalledWith('sess-1', 'wa-msg-1', 'edited');
+      expect(messageProjector.recordOutboundMessageEdit).toHaveBeenCalledWith('sess-1', 'wa-msg-1', 'edited');
       expect(repository.update).not.toHaveBeenCalled();
       expect(res).toEqual({ messageId: 'wa-msg-1', timestamp: 1706868000 });
     });
@@ -1306,7 +1311,7 @@ describe('MessageService', () => {
       await expect(
         service.editMessage('sess-1', { chatId: 'test@c.us', messageId: 'wa-msg-1', body: 'edited' }),
       ).rejects.toBeInstanceOf(NotFoundException);
-      expect(sessionService.recordOutboundMessageEdit).not.toHaveBeenCalled();
+      expect(messageProjector.recordOutboundMessageEdit).not.toHaveBeenCalled();
     });
 
     it('throws BadRequestException when the session is not started', async () => {
@@ -1337,7 +1342,7 @@ describe('MessageService', () => {
       ).rejects.toThrow('Message sending blocked by plugin');
 
       expect(mockEngine.editMessage).not.toHaveBeenCalled();
-      expect(sessionService.recordOutboundMessageEdit).not.toHaveBeenCalled();
+      expect(messageProjector.recordOutboundMessageEdit).not.toHaveBeenCalled();
     });
 
     it('threads a plugin-rewritten edit body through to the engine and the stored row', async () => {
@@ -1349,7 +1354,7 @@ describe('MessageService', () => {
       await service.editMessage('sess-1', { chatId: 'test@c.us', messageId: 'wa-msg-1', body: 'secret' });
 
       expect(mockEngine.editMessage).toHaveBeenCalledWith('test@c.us', 'wa-msg-1', 'redacted');
-      expect(sessionService.recordOutboundMessageEdit).toHaveBeenCalledWith('sess-1', 'wa-msg-1', 'redacted');
+      expect(messageProjector.recordOutboundMessageEdit).toHaveBeenCalledWith('sess-1', 'wa-msg-1', 'redacted');
     });
   });
