+++ src/modules/session/session.module.ts
@@ -5,6 +5,7 @@ import { Message } from '../message/entities/message.entity';
 import { SessionService } from './session.service';
 import { SessionLidResolver } from './session-lid-resolver.service';
 import { SessionLivenessWatchdog } from './session-liveness-watchdog.service';
+import { MessageProjector } from './message-projector.service';
 import { SessionController } from './session.controller';
 import { WebhookModule } from '../webhook/webhook.module';
 import { StatusStoreModule } from '../status-store/status-store.module';
@@ -14,7 +15,7 @@ import { StatusStoreModule } from '../status-store/status-store.module';
   // one-directional — no forwardRef() needed.
   imports: [TypeOrmModule.forFeature([Session, Message], 'data'), WebhookModule, StatusStoreModule],
   controllers: [SessionController],
-  providers: [SessionService, SessionLidResolver, SessionLivenessWatchdog],
-  exports: [SessionService],
+  providers: [SessionService, SessionLidResolver, SessionLivenessWatchdog, MessageProjector],
+  exports: [SessionService, MessageProjector],
 })
 export class SessionModule {}
